+++ flow/envs/multiagent/__init__.py
@@ -4,9 +4,9 @@ from flow.envs.multiagent.base import MultiEnv
 from flow.envs.multiagent.ring.wave_attenuation import \
     MultiWaveAttenuationPOEnv
 
-from flow.envs.multiagent.ring.accel import MultiAgentAccelEnv
+from flow.envs.multiagent.ring.accel import AdversarialAccelEnv
 from flow.envs.multiagent.traffic_light_grid import MultiTrafficLightGridPOEnv
 from flow.envs.multiagent.highway import MultiAgentHighwayPOEnv
 
-__all__ = ['MultiEnv', 'MultiAgentAccelEnv', 'MultiWaveAttenuationPOEnv',
+__all__ = ['MultiEnv', 'AdversarialAccelEnv', 'MultiWaveAttenuationPOEnv',
            'MultiTrafficLightGridPOEnv', 'MultiAgentHighwayPOEnv']
