+++ flow/multiagent_envs/__init__.py
@@ -3,13 +3,13 @@
 from flow.multiagent_envs.multiagent_env import MultiEnv
 from flow.multiagent_envs.loop.wave_attenuation import \
     MultiWaveAttenuationPOEnv
-from flow.multiagent_envs.loop.loop_accel import MultiAgentAccelEnv
+from flow.multiagent_envs.loop.loop_accel import AdversarialAccelEnv
 from flow.multiagent_envs.traffic_light_grid import MultiTrafficLightGridPOEnv
 from flow.multiagent_envs.highway import MultiAgentHighwayPOEnv
 
 __all__ = [
     'MultiEnv',
-    'MultiAgentAccelEnv',
+    'AdversarialAccelEnv',
     'MultiWaveAttenuationPOEnv',
     'MultiTrafficLightGridPOEnv',
     'MultiAgentHighwayPOEnv'
